:building_construction: update inviteGameRoom
