+++ frontend/src/components/chat/MenuChat.tsx
@@ -2,11 +2,11 @@ import { useEffect, useState } from "react";
 import { shallowEqual, useDispatch, useSelector } from "react-redux";
 import { socket } from "../../socket/socket";
 import { RootState } from "../../redux/rootReducer";
-import { chatRoom, ChatData, InputChatRoom, ChatUser } from "../../types/chatTypes";
+import { updateChat } from "../../redux/chatReducer";
+import { chatRoom, ChatData } from "../../types/chatTypes";
 import MenuChatBox from "./MenuChatBox";
 import AddChatModal from "../modals/AddChatModal";
 import PublicChatModal from "../modals/PublicChatModal"
-import { updateChat } from "../../redux/chatReducer";
 
 export default function MenuChat(){
 	const dispatch = useDispatch();
@@ -31,34 +31,9 @@ export default function MenuChat(){
 			setRoom(data);
 			dispatch(updateChat(data));
 		});
-		socket.on("updateChatRoom", (data:InputChatRoom)=>{
-			console.log("update Chat Room!");
-			console.log(data, room);
-			let tmp:ChatData = chatroom;
-			const idx = tmp.order.indexOf(data.chatid);
-			if (data.title){ tmp.chatroom[idx].title = data.title; }
-			if (data.type){ tmp.chatroom[idx].type = data.type; }
-			if (data.addManager){ data.addManager.forEach((man:string)=>{
-				let index:number = -1;
-				index = tmp.chatroom[idx].manager?.findIndex((p:string)=>p === man);
-				if (index === -1){ tmp.chatroom[idx].manager.push(man); }
-			}); }
-			if (data.deleteManager){ data.deleteManager.forEach(man=>tmp.chatroom[idx].manager = tmp.chatroom[idx].manager?.filter((person: string)=> man !== person)); }
-			if (data.enterUser){ data.enterUser.forEach((user:ChatUser)=>{
-				let index:number = -1;
-				index = tmp.chatroom[idx].members.findIndex((p:ChatUser)=>p.userid===user.userid);
-				if (index === -1){ tmp.chatroom[idx].members.push(user); }
-			}); }
-			if (data.exitUser){ data.exitUser.forEach(user=>tmp.chatroom[idx].members = tmp.chatroom[idx].members?.filter((person:ChatUser)=> user !== person.userid)); }
-			if (data.switchOwner){ tmp.chatroom[idx].owner = data.switchOwner; }
-			setRoom({...tmp});
-			dispatch(updateChat(tmp));
-		});
-
 		return ()=>{
 			socket.off("enterChatRoom");
 			socket.off("myChatRoom");
-			socket.off("updateChatRoom");
 		}
 	}, [chatroom, dispatch, room]);
